Day 3/requirements.txt
tensorflow-cpu==2.3.1
Flask==1.1.2